refactor: extract seller_wantlist_handoff_enabled? into shared concern
