+++ app/controllers/application_controller.rb
@@ -1,3 +1,5 @@
 class ApplicationController < ActionController::Base
+  include SellerWantlistHandoff
+
   allow_browser versions: :modern
 end
