+++ app/controllers/pile_controller.rb
@@ -35,14 +35,4 @@ class PileController < ApplicationController
     end
   end
 
-  private
-
-  def seller_wantlist_handoff_enabled?
-    return false unless Settings.respond_to?(:features)
-    return false unless Settings.features.respond_to?(:seller_wantlist_handoff)
-
-    ActiveModel::Type::Boolean.new.cast(
-      ENV.fetch("SELLER_WANTLIST_HANDOFF_ENABLED", Settings.features.seller_wantlist_handoff.enabled.to_s)
-    )
-  end
 end
